+++ module-system/module-implementations/sov-attester-incentives/src/tests/attestation_processing.rs
@@ -1,3 +1,4 @@
+use sov_bank::GAS_TOKEN_ID;
 use sov_modules_api::optimistic::Attestation;
 use sov_modules_api::{Context, WorkingSet};
 use sov_modules_core::GasMeter;
@@ -5,7 +6,7 @@ use sov_prover_storage_manager::new_orphan_storage;
 
 use crate::call::AttesterIncentiveErrors;
 use crate::tests::helpers::{
-    execution_simulation, setup, BOND_AMOUNT, INITIAL_BOND_AMOUNT, INIT_HEIGHT,
+    setup, ExecutionSimulationVars, BOND_AMOUNT, INITIAL_BOND_AMOUNT, INIT_HEIGHT,
 };
 type S = sov_test_utils::TestSpec;
 
@@ -15,7 +16,7 @@ fn test_process_valid_attestation() {
     let tmpdir = tempfile::tempdir().unwrap();
     let storage = new_orphan_storage(tmpdir.path()).unwrap();
     let working_set = WorkingSet::new(storage.clone());
-    let (module, token_id, attester_address, _, sequencer, mut working_set) = setup(working_set);
+    let (module, attester_address, _, sequencer, mut working_set) = setup(working_set);
 
     // Assert that the attester has the correct bond amount before processing the proof
     assert_eq!(
@@ -29,8 +30,14 @@ fn test_process_valid_attestation() {
     // Simulate the execution of a chain, with the genesis hash and two transitions after.
     // Update the chain_state module and the optimistic module accordingly
     let state_checkpoint = working_set.checkpoint().0;
-    let (mut exec_vars, state_checkpoint) =
-        execution_simulation(3, &module, &storage, attester_address, state_checkpoint);
+    let (mut exec_vars, state_checkpoint) = ExecutionSimulationVars::execute(
+        3,
+        &module,
+        &storage,
+        &sequencer,
+        &attester_address,
+        state_checkpoint,
+    );
 
     let context = Context::<S>::new(attester_address, sequencer, 1);
 
@@ -91,10 +98,13 @@ fn test_process_valid_attestation() {
     assert_eq!(
         module
             .bank
-            .get_balance_of(attester_address, token_id, &mut working_set)
+            .get_balance_of(attester_address, GAS_TOKEN_ID, &mut working_set)
             .unwrap(),
         // The attester is bonded at the beginning so he loses BOND_AMOUNT
-        INITIAL_BOND_AMOUNT - BOND_AMOUNT + 2 * BOND_AMOUNT
+        INITIAL_BOND_AMOUNT - BOND_AMOUNT
+            + 2 * module
+                .burn_rate()
+                .apply(ExecutionSimulationVars::tx_reward())
     );
 }
 
@@ -103,7 +113,7 @@ fn test_burn_on_invalid_attestation() {
     let tmpdir = tempfile::tempdir().unwrap();
     let storage = new_orphan_storage(tmpdir.path()).unwrap();
     let working_set = WorkingSet::new(storage.clone());
-    let (module, _token_id, attester_address, _, sequencer, mut working_set) = setup(working_set);
+    let (module, attester_address, _, sequencer, mut working_set) = setup(working_set);
 
     // Assert that the prover has the correct bond amount before processing the proof
     assert_eq!(
@@ -120,8 +130,14 @@ fn test_burn_on_invalid_attestation() {
     // Simulate the execution of a chain, with the genesis hash and two transitions after.
     // Update the chain_state module and the optimistic module accordingly
     let state_checkpoint = working_set.checkpoint().0;
-    let (mut exec_vars, state_checkpoint) =
-        execution_simulation(3, &module, &storage, attester_address, state_checkpoint);
+    let (mut exec_vars, state_checkpoint) = ExecutionSimulationVars::execute(
+        3,
+        &module,
+        &storage,
+        &sequencer,
+        &attester_address,
+        state_checkpoint,
+    );
 
     let transition_2 = exec_vars.pop().unwrap();
     let transition_1 = exec_vars.pop().unwrap();