+++ module-system/module-implementations/sov-attester-incentives/src/tests/invariant.rs
@@ -5,7 +5,7 @@ use sov_prover_storage_manager::new_orphan_storage;
 
 use crate::call::AttesterIncentiveErrors;
 use crate::tests::helpers::{
-    execution_simulation, setup, BOND_AMOUNT, DEFAULT_ROLLUP_FINALITY, INIT_HEIGHT,
+    setup, ExecutionSimulationVars, BOND_AMOUNT, DEFAULT_ROLLUP_FINALITY, INIT_HEIGHT,
 };
 type S = sov_test_utils::TestSpec;
 
@@ -15,7 +15,7 @@ fn test_transition_invariant() {
     let tmpdir = tempfile::tempdir().unwrap();
     let storage = new_orphan_storage(tmpdir.path()).unwrap();
     let working_set = WorkingSet::new(storage.clone());
-    let (module, _token_id, attester_address, _, sequencer, mut working_set) = setup(working_set);
+    let (module, attester_address, _, sequencer, mut working_set) = setup(working_set);
 
     // Assert that the attester has the correct bond amount before processing the proof
     assert_eq!(
@@ -32,8 +32,14 @@ fn test_transition_invariant() {
     // Simulate the execution of a chain, with the genesis hash and two transitions after.
     // Update the chain_state module and the optimistic module accordingly
     let state_checkpoint = working_set.checkpoint().0;
-    let (exec_vars, state_checkpoint) =
-        execution_simulation(20, &module, &storage, attester_address, state_checkpoint);
+    let (exec_vars, state_checkpoint) = ExecutionSimulationVars::execute(
+        20,
+        &module,
+        &storage,
+        &sequencer,
+        &attester_address,
+        state_checkpoint,
+    );
     let mut working_set = state_checkpoint.to_revertable(GasMeter::unmetered());
 
     let context = Context::<S>::new(attester_address, sequencer, INIT_HEIGHT + 2);