+++ module-system/module-implementations/sov-prover-incentives/src/genesis.rs
@@ -1,6 +1,6 @@
 use anyhow::Result;
+use serde::de::DeserializeOwned;
 use serde::{Deserialize, Serialize};
-use sov_bank::TokenId;
 use sov_modules_api::{DaSpec, Spec, WorkingSet, Zkvm};
 
 use crate::{Amount, ProverIncentives};
@@ -10,13 +10,10 @@ use crate::{Amount, ProverIncentives};
 /// the allowed verifier method and a set of initial provers with their
 /// bonding amount.
 #[derive(Debug, Clone, Serialize, Deserialize)]
+#[serde(bound = "S::Address: Serialize + DeserializeOwned")]
 pub struct ProverIncentivesConfig<S: sov_modules_api::Spec> {
-    /// The address of the account holding the reward token supply
-    pub reward_token_supply_address: S::Address,
     /// A penalty for provers who submit a proof for transitions that were already proven
     pub proving_penalty: Amount,
-    /// The ID of the token to be used for bonding.
-    pub bonding_token_id: TokenId,
     /// The minimum bond for a prover.
     pub minimum_bond: u64,
     /// A code commitment to be used for verifying proofs
@@ -42,10 +39,6 @@ impl<S: sov_modules_api::Spec, Da: DaSpec> ProverIncentives<S, Da> {
         self.minimum_bond.set(&config.minimum_bond, working_set);
         self.commitment_of_allowed_verifier_method
             .set(&config.commitment_of_allowed_verifier_method, working_set);
-        self.bonding_token_id
-            .set(&config.bonding_token_id, working_set);
-        self.reward_token_supply_address
-            .set(&config.reward_token_supply_address, working_set);
         self.proving_penalty
             .set(&config.proving_penalty, working_set);
         self.last_claimed_reward.set(&0, working_set);